Rename Reader,Writer to CheckpoointReader,Writer move to own headers
for PR #111P
